add GoogleAuthService class
